requirements.txt
alembic==1.3.1
atomicwrites==1.3.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
cachetools==4.0.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.1
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.0
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-language==1.3.0
googleapis-common-protos==1.51.0
grpcio==1.26.0
httplib2==0.17.0
idna==2.8
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
kiwisolver==1.1.0
lxml==4.4.1
Mako==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
more-itertools==7.2.0
mysqlclient==1.4.4
nbconvert==5.6.1
nbformat==5.0.3
nltk==3.4.5
notebook==6.0.2
numpy==1.17.4
packaging==19.2
pandas==0.25.3
pandoc==1.0.2
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
pluggy==0.13.0
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.11.2
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.4.2
pyparsing==2.4.5
pyrsistent==0.15.7
pytest==5.2.4
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.3
pywinpty==0.5.7
PyYAML==5.3
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
rsa==4.0
saltools==0.1.9
Send2Trash==1.5.0
six==1.12.0
SQLAlchemy==1.3.8
SQLAlchemy-Utils==0.35.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.2
uritemplate==3.0.1
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
zipp==0.6.0
